Enable printf to Serial
This simplifies debug printing in .c files, which do not have access to
the Serial C++ object. To use this, just include stdio.h and call
printf. This replaces the `debug()` and `debug_str()` functions that
were previously present.

This support can be disabled in config.h, which saves some flash space
(but note that this also makes failed assertions silent!).