Correct power pin documentations
This previously referred to SX1272 pins that are used as regulator
outputs, so powering those would actually be harmful. This resulted from
the fact that the SX1272 evaluation board has multiple power pins, but
it turns out these do not correspond to the SX1272 pins as I had
previously assumed.

The documentation now explicitely mentions the evaluation board pin
names, which should be more clear.