Only print the length of received data in example
Previously, the raw data was printed, but since that is likely binary
data, this just prints the length. This should be sufficient to let
people know there is data and then write proper decoding for it.
Printing in hex would be useful, but printing an array of bytes in hex
is a bit cumbersome in Arduino.