Put printf format strings in flash on AVR
This adds a lmic_printf macro that adds PSTR() around the format string
argument to put it into flash, and calls printf_P() instead of printf()
to get the format string from flash while printing. This greatly reduces
the RAM usage when LMIC_DEBUG_LEVEL is > 0 (and does not affect the
generated code otherwise).