Fix compilation with LMIC_DEBUG_LEVEL > 1
I added a variable to assist in printing the deadline for scheduled
jobs, but then decided to leave out the variable. Apparently, I left one
use of this variable in, breaking the build. Looking at it again, it's
probably best to keep the variable, though it's now renamed to be a bit
more logical.

This fixes #114.