Rename the trace functions LMICOS_logEvent() etc.
Renaming makes sense, because from the LMIC's point of view, logging
functions are provided by the OS/HAL. It might make further sense to
move the LOG function from the compliance test script to the OS/HAL,
so the logging can be used by any script without makeing them recreate
the facility. On the other hand, that would make it more of a headache
to override the default in the normal case.