Fix #453: US v1.3 compliance 12.5 DL count hack
The US v1.3 compliance test section 12.5 requires that we not
increment the downlink count for an EchoRequest unless the
EchoResponse is permitted with the current data rate (what the
LMIC code calls "feasibility"). We can't tell if a message is
feasible unless we actually try to send it, and we increment the
downlink count centrally. Workaournd is to decrement the downlink
count if the LMIC says the EchoResponse is infeasible.